requirements.txt
# Python module requirements
pyyaml
requests
python-dateutil
